Render minecraft avatar's by ourself, instead of letting minotar.net do it.
That's great for privacy, because mintar.net uses Google Analytics and I'm a bit paranoid ;) .
Every avatar is first loaded from the original minecraft skin server (which become's converted to the head), and then cached by username and size for the next requests.
I also included a custom twig extension, which makes it much easier to get an avatar url/tag.