use MopaBootstrapBundle
In order to require the newest knp-menu, I had to remove the
SonataAdminBundle, which required an old version.
To not to add even more diffs to this commit, I decided to add the user
context later.
Furthermore, I had problems with the assets:install command trying to
remove the bootstrap symlink, so I removed the script from
composer.json, which might fail some bundles.